warning

+++ stepmania/src/ScreenNetEvaluation.cpp
@@ -172,7 +172,7 @@ void ScreenNetEvaluation::TweenOffScreen( )
 
 void ScreenNetEvaluation::UpdateStats()
 {
-	if ( m_iCurrentPlayer >= NSMAN->m_EvalPlayerData.size() )
+	if( m_iCurrentPlayer >= (int) NSMAN->m_EvalPlayerData.size() )
 		return;
 
 	m_Grades[m_pActivePlayer].SetGrade( (PlayerNumber)m_pActivePlayer, (Grade)NSMAN->m_EvalPlayerData[m_iCurrentPlayer].grade );
